apps/backend/requirements.txt
uvicorn
fastapi
pyyaml
httpx
psutil